requirements.txt
# Core dependencies
streamlit>=1.25.0
requests>=2.28.0
numpy>=1.23.0
moviepy>=1.0.3
torchaudio>=2.0.0
torch>=2.0.0
transformers>=4.28.0
scipy>=1.10.0
matplotlib>=3.6.0

# Audio processing
librosa>=0.10.0
SoundFile>=0.10.0
PySoundFile>=0.9.0

# Speech recognition
openai-whisper>=20230314
